conf/machine: fix KERNEL_ROOTSPEC_DEFAULT for SDcard-based Nanos
When adding support for booting from external devices, the
composition of the KERNEL_ROOTSPEC setting got refactored so
that the Nano and 2GB Nano dev kits no longer needed to override
the full setting. I forgot about that when updating the Nano
machine configs, and just copy-pasted the old setting to
the new KERNEL_ROOTSPEC_DEFAULT variable. Fix that so the
setting includes just the device name, like it's supposed to.

Signed-off-by: Matt Madison <matt@madison.systems>